% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseJaitinSpleen}
\alias{MouseJaitinSpleen}
\title{Mouse Spleen single cell RNA-seq data due to Jaitin et al (2014)}
\format{Data is an \code{\link{ExpressionSet}} object. The
\code{exprs} slot contains the gene read counts in a gene
by sample matrix with 20190 genes and 4590 samples.
\code{phenoData} slot contains meta information of the samples on 15 features
including batches, sorting markers, ERCC information etc.\code{featureData}
slot contains row information which are usually ensemebl gene ID or gene
symbols.}
\usage{
MouseJaitinSpleen
}
\description{
A dataset containing single cell RNA-seq data from mouse spleen due to Jaitin et al (2014)
}
\examples{
data(MouseJaitinSpleen)
exprs(MouseJaitinSpleen)
pData(MouseJaitinSpleen)
featureNames(MouseJaitinSpleen)
}
\references{
Diego Adhemar Jaitin, Ephraim Kenigsberg, Hadas Keren-Shaul, Naama Elefant, Franziska Paul, Irina Zaretsky, Alexander Mildner, Nadav Cohen, Steffen Jung, Amos Tanay, Ido Amit. Massively Parallel Single-Cell RNA-Seq for Marker-Free Decomposition of Tissues into Cell Types. Science 343, 776-779 (2014) DOI: 10.1126/science.1247651
}
\keyword{datasets}

